move into ValueObject namespace
