+++ rules/CodingStyle/ClassNameImport/UsedImportsResolver.php
@@ -8,6 +8,7 @@ use PhpParser\Node\Identifier;
 use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\UseUse;
+use Rector\CodingStyle\ClassNameImport\ValueObject\UsedImports;
 use Rector\Core\PhpParser\Node\BetterNodeFinder;
 use Rector\NodeNameResolver\NodeNameResolver;
 use Rector\StaticTypeMapper\ValueObject\Type\AliasedObjectType;
