+++ rules/CodingStyle/ClassNameImport/ValueObject/UsedImports.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\CodingStyle\ClassNameImport;
+namespace Rector\CodingStyle\ClassNameImport\ValueObject;
 
 use Rector\StaticTypeMapper\ValueObject\Type\AliasedObjectType;
 use Rector\StaticTypeMapper\ValueObject\Type\FullyQualifiedObjectType;
